+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 .DS_store
 .cache
 dist
+/Users/angelfernandez/Coding/TKH/2020_Innovation_Fellowship/Progressive_Web_Development/Phase_1_Project/-Munch-Map-Redux/src/js/API/gmaps.js
