+++ src/index.html
@@ -1,20 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Munch Map</title>
-		<link rel="stylesheet" href="../src/css/styles.css" />
-		<script src="https://polyfill.io/v3/polyfill.min.js?features=default"></script>
-	</head>
-	<body>
-		<h1>Test</h1>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Munch Map</title>
+    <link rel="stylesheet" href="../src/css/styles.css" />
+    <script src="https://polyfill.io/v3/polyfill.min.js?features=default"></script>
+  </head>
+  <body>
+    <h1>Test</h1>
 
-		<div id="restaurants"></div>
-		<div id="map"></div>
-		<!-- <script src="../src/js/API/gmaps.js"></script>
+    <div id="restaurants"></div>
+    <div id="map"></div>
+    <!-- <script src="../src/js/API/gmaps.js"></script>
 		<script src="../src/js/API/zomato.js"></script> -->
-		<script src="../src/js/test.js"></script>
-		<!-- <script src="../src/js/API/gmaps.js"></script> -->
-	</body>
+    <!-- <script src="../src/js/test.js"></script> -->
+    <!-- <script src="../src/js/API/gmaps.js"></script> -->
+    <script src="../src/js/API/zomato-gmap.js"></script>
+  </body>
 </html>
